Missing changes of #980 (#986)
* Add Readme, simplify generaly chain data types.
